requirements.txt
aniso8601==8.1.0
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.25.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2020.4
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
